refactor: ULIDGenerator package
