+++ src/main/java/com/gaaji/townlife/service/domain/category/Category.java
@@ -23,7 +23,7 @@ import java.util.Optional;
 })
 public class Category {
     @Id
-    @GenericGenerator(name = "ulidGenerator", strategy = "com.gaaji.townlife.global.utils.ULIDGenerator")
+    @GenericGenerator(name = "ulidGenerator", strategy = "com.gaaji.townlife.global.utils.ulid.ULIDGenerator")
     @GeneratedValue(generator = "ulidGenerator")
     private String id;
     @Column(unique = true)
