+++ src/main/java/com/gaaji/townlife/service/domain/comment/Comment.java
@@ -22,7 +22,7 @@ import java.util.List;
 })
 public abstract class Comment extends BaseEntity {
     @Id
-    @GenericGenerator(name = "ulidGenerator", strategy = "com.gaaji.townlife.global.utils.ULIDGenerator")
+    @GenericGenerator(name = "ulidGenerator", strategy = "com.gaaji.townlife.global.utils.ulid.ULIDGenerator")
     @GeneratedValue(generator = "ulidGenerator")
     private String id;
     private String userId;
