+++ src/main/java/com/gaaji/townlife/service/domain/townlife/AttachedImage.java
@@ -13,7 +13,7 @@ import javax.persistence.*;
 @ToString
 public class AttachedImage {
     @Id
-    @GenericGenerator(name = "ulidGenerator", strategy = "com.gaaji.townlife.global.utils.ULIDGenerator")
+    @GenericGenerator(name = "ulidGenerator", strategy = "com.gaaji.townlife.global.utils.ulid.ULIDGenerator")
     @GeneratedValue(generator = "ulidGenerator")
     private String id;
     private int orderIndex;
